+++ example/lib/pages/search_bar_page.dart
@@ -1,8 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_widgetz/flutter_widgetz.dart';
 
-class SearchBarPage extends StatelessWidget {
+class SearchBarPage extends StatefulWidget {
   const SearchBarPage({Key? key}) : super(key: key);
+
+  @override
+  State<SearchBarPage> createState() => _SearchBarPageState();
+}
+
+class _SearchBarPageState extends State<SearchBarPage> {
+  late bool _isVisible;
+
+  @override
+  void initState() {
+    super.initState();
+    _isVisible = true;
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -19,6 +32,18 @@ class SearchBarPage extends StatelessWidget {
             onChanged: print,
             onClear: () {},
           ),
+          CustomElevatedButton(
+            text: 'Toggle Visibility',
+            onPressed: () => setState(() {
+              _isVisible = !_isVisible;
+            }),
+          ),
+          SearchBar(
+            labelText: 'Search Bar Three',
+            onChanged: print,
+            isVisible: _isVisible,
+          ),
+          const Text('Some other widgets....'),
         ],
       ),
     );