+++ lib/src/search_bar.dart
@@ -8,20 +8,29 @@ class SearchBar extends StatefulWidget {
   const SearchBar({
     Key? key,
     required this.onChanged,
+    this.alignment = Alignment.topCenter,
     this.autofocus = false,
+    this.duration = kThemeChangeDuration,
     this.isVisible = true,
     this.keyboardType = TextInputType.text,
     this.labelText = 'Search',
     this.onClear,
     this.padding = EdgeInsets.zero,
+    this.prefixIcon = Icons.search,
     this.value = '',
   }) : super(key: key);
 
   /// Called when the value of the search bar changes.
   final ValueChanged<String> onChanged;
 
+  /// The alignment of the child within the parent.
+  final Alignment alignment;
+
   /// Determines whether to auto focus this input.
   final bool autofocus;
+
+  /// The duration when transitioning this widget's size.
+  final Duration duration;
 
   /// Determines whether to show this input.
   final bool isVisible;
@@ -37,6 +46,9 @@ class SearchBar extends StatefulWidget {
 
   /// The amount of space by which to inset the child.
   final EdgeInsets padding;
+
+  /// An icon that appears before the editable part of the text field.
+  final IconData prefixIcon;
 
   /// The text value of this input.
   final String value;
@@ -63,24 +75,29 @@ class _SearchBarState extends State<SearchBar> {
 
   @override
   Widget build(BuildContext context) {
-    if (!widget.isVisible) {
-      return const SizedBox();
-    }
-
-    return Padding(
-      padding: widget.padding,
-      child: TextFormField(
-        autofocus: widget.autofocus,
-        controller: _controller,
-        keyboardType: widget.keyboardType,
-        textInputAction: TextInputAction.search,
-        decoration: InputDecoration(
-          labelText: widget.labelText,
-          prefixIcon: const Icon(Icons.search),
-          suffixIcon: _getSuffix(),
-        ),
-        onChanged: widget.onChanged,
-      ),
+    return AnimatedSize(
+      alignment: widget.alignment,
+      duration: widget.duration == Duration.zero
+          // zero duration causes an error
+          ? const Duration(milliseconds: 1)
+          : widget.duration,
+      child: !widget.isVisible
+          ? const SizedBox()
+          : Padding(
+              padding: widget.padding,
+              child: TextFormField(
+                autofocus: widget.autofocus,
+                controller: _controller,
+                keyboardType: widget.keyboardType,
+                textInputAction: TextInputAction.search,
+                decoration: InputDecoration(
+                  labelText: widget.labelText,
+                  prefixIcon: Icon(widget.prefixIcon),
+                  suffixIcon: _getSuffix(),
+                ),
+                onChanged: widget.onChanged,
+              ),
+            ),
     );
   }